+++ src/commonApi.ts
@@ -50,7 +50,12 @@ export abstract class CommonApi<TMessage, TRequestBody> {
 	/** System prompts to include in requests. */
 	protected _systemContent: string | undefined;
 
-	constructor() {}
+	/** Set the model ID for logging purposes. */
+	protected _modelId = "";
+
+	constructor(modelId: string) {
+		this._modelId = modelId;
+	}
 
 	/**
 	 * Convert VS Code chat messages to specific api message format.